Add core event handling and synchronization features
- Introduced a new child handler for processing event messages and managing workflow events.
- Implemented an abstract chain class for handling blockchain interactions, specifically for Ethereum Sepolia.
- Created an event handler factory to build specific chain handlers based on event configurations.
- Added utility functions for fetching active workflows and managing Redis synchronization.
- Updated existing files for improved module imports and consistency in naming conventions.
- Added a .gitignore file to exclude sensitive files and directories from version control.
