+++ sc-event-tracker/.gitignore
@@ -1,0 +1,5 @@
+.env
+
+./worker/.env
+
+node_modules/
