+++ sc-event-tracker/src/core/synchronization/container-down-action.js
