+++ sc-event-worker/src/config/logger.js
@@ -1,9 +1,9 @@
 class Logger {
   /**
-   * @param {console} logger - The logger object to use.
+   * @param {console} loggerInstance - The logger object to use.
    */
-  constructor(logger) {
-    this.logger = logger;
+  constructor(loggerInstance) {
+    this.logger = loggerInstance;
   }
 
   /**
